refactor: rename database package to db for consistency
